one pipe harness: the integration tests stop carrying five folk copies
The pipe-family machinery existed per test -- frame
builders, recv loops, readiness waits, journal readers --
with drifting deadlines, path prefixes (the
doubled-backslash class lived in exactly these copies), and
header parses (the version-field bug was fixed in one file
only). rpipe.py is the one copy: frame/recv against the RTRD
header, open_pipe (the connection) and wait_pipe (readiness)
split by shape, wait_sock dual-mode for the tests that also
run POSIX-side over UDS, journal_records, run, taskkill.
The seven pipe-family tests import it; the POSIX-only
wait_sock users keep their socket-shaped copies.
